+++ ruby/lib/essence/components/link.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class Essence::Link < Essence::Essence
-  CLASS = "inline-flex w-fit items-center gap-1 font-medium text-gray-900 border-b-2 hover:border-gray-900 transition-colors"
+  BASE = "inline-flex w-fit items-center gap-1 font-medium text-gray-900 border-b-2 hover:border-gray-900 transition-colors"
   KINDS = {
     regular: "border-transparent",
     underline: "border-gray-200"
@@ -11,7 +11,7 @@ class Essence::Link < Essence::Essence
 
   def initialize(kind: :regular, **attributes)
     @attributes = attributes
-    @attributes[:class] = @attributes[:class] ? TAILWIND_MERGER.merge([ CLASS, @attributes[:class] ]) : CLASS
+    @attributes[:class] = @attributes[:class] ? TAILWIND_MERGER.merge([ BASE, @attributes[:class] ]) : BASE
   end
 
   def view_template(&)